patches: Prepare to use 'gbp pq'
We don't currently need to carry any patches, but when the need
will arise again in the future we want to use 'gbp pq' instead
of plain quilt to manage them.

Set the stage by configuring gbp accordingly and removing all
custom scripting around patches. We also get to remove the .pc
directory, which really should never have been committed to git
in the first place.